refactor: move Thunderstore mods view into dedicated folder
